git-svn-id: https://gamescrafters.svn.sourceforge.net/svnroot/gamescrafters/GamesmanJava@4437 a668be5d-ef11-cb4c-8bfd-e08cbb24f2ba

+++ src/edu/berkeley/gamesman/master/LocalMaster.java
@@ -137,8 +137,7 @@ public final class LocalMaster implements Master, TaskFactory {
 		@Override
 		public void update() {
 			long elapsedMillis = System.currentTimeMillis() - start;
-			double thousandpct = completed.doubleValue()
-					/ (total.doubleValue() / 100000);
+			double thousandpct = completed / (total / 100000D);
 			double pct = thousandpct / 1000;
 			long totalMillis = (long) ((double) elapsedMillis * 100 / pct);
 			System.out.print("Task: " + name + ", "
